recommend(q), recommendtest(), bm25(), avg_leng()
recommend() recommends a specific tweet chosen by the user and displays more tweets like it. This would have been done through a "like" feature. recommendtest() does a deep dive into tweets, it seems as though the dataset is clustered in a way when using this function as many times i get stuck on tweets with the word "hello". bm25(), an implementation of the bm25 method that was presented to us in class, the method may not work some times but does not cause any fatal errors when running or failing.