Final Version with a the ability to like a post
Using like() with the number of what post you likes will add posts similar to the one you liked to your list. Added a Euclidean distance formula and have a recommendation feature based on the Euclidean distance. Was unable to make a stable version that supported a mix of Euclidean and cosine similarity.